+++ CHANGELOG.md
@@ -8,25 +8,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- ...
+
+- Wrapper for training a gensim word2vec model [#13](https://github.com/iomega/spec2vec/tree/13-gensim-wrapper)
 
 ### Changed
+
 - Extend spec2vec similarity calculation to handle missing words [#9](https://github.com/iomega/spec2vec/issues/9)
 - Updated the integration test [#7](https://github.com/iomega/spec2vec/issues/7)
 
 ### Fixed
+
 - ...
 
 ### Removed
+
 - ...
 
 ## [0.1.0] - 2020-06-02
 
 ### Added
+
 - Matchms as dependency [#4](https://github.com/iomega/spec2vec/pull/4)
 - Bump2version config
 
 ### Changed
+
 - Splitted spec2vec from [matchms]. See (https://github.com/matchms/matchms) [#1](https://github.com/iomega/spec2vec/pull/1) [#4](https://github.com/iomega/spec2vec/pull/4)
   - Updated packaging related configuration
   - Update the GH Actions workflows
@@ -38,9 +44,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Removed
+
 - Fossa configuration
 - Flowchart
-
 
 [Unreleased]: https://github.com/iomega/spec2vec/compare/0.1.0...HEAD
 [0.1.0]: https://github.com/iomega/spec2vec/releases/tag/0.1.0